+++ CumulusMX/Stations/GW1000Station.cs
@@ -1432,7 +1432,7 @@ namespace CumulusMX
 
 						if (cumulus.StationOptions.CalculateSLP)
 						{
-							var slp = MeteoLib.GetSeaLevelPressure(AltitudeM(cumulus.Altitude), ConvertUnits.UserPressToMB(StationPressure), ConvertUnits.UserTempToC(OutdoorTemperature), cumulus.Latitude);
+							var slp = MeteoLib.GetSeaLevelPressure(ConvertUnits.AltitudeM(cumulus.Altitude), ConvertUnits.UserPressToMB(StationPressure), ConvertUnits.UserTempToC(OutdoorTemperature), cumulus.Latitude);
 							DoPressure(ConvertUnits.PressMBToUser(slp), dateTime);
 						}
 					}