+++ CumulusMX/Stations/TempestStation.cs
@@ -174,7 +174,7 @@ namespace CumulusMX
 
 				// Pressure =============================================================
 				DoStationPressure(ConvertUnits.PressMBToUser((double) historydata.StationPressure));
-				DoPressure(ConvertUnits.PressMBToUser(MeteoLib.GetSeaLevelPressure(AltitudeM(cumulus.Altitude), ConvertUnits.UserPressToHpa(StationPressure), (double) historydata.Temperature)), timestamp);
+				DoPressure(ConvertUnits.PressMBToUser(MeteoLib.GetSeaLevelPressure(ConvertUnits.AltitudeM(cumulus.Altitude), ConvertUnits.UserPressToHpa(StationPressure), (double) historydata.Temperature)), timestamp);
 
 				// Outdoor Humidity =====================================================
 				DoOutdoorHumidity((int) historydata.Humidity, timestamp);
@@ -330,7 +330,7 @@ namespace CumulusMX
 							ConvertUnits.WindMSToUser((double) wp.Observation.WindAverage),
 							ts);
 
-						var alt = AltitudeM(cumulus.Altitude);
+						var alt = ConvertUnits.AltitudeM(cumulus.Altitude);
 						var seaLevel = MeteoLib.GetSeaLevelPressure(alt, (double) wp.Observation.StationPressure,
 							(double) wp.Observation.Temperature);
 						DoPressure(ConvertUnits.PressMBToUser(seaLevel), ts);